No associations on results from the store
I think this might make stuff "flatter" in the long run. The Store has to handle
more logic that way, but I think it will pay off.
This will allow me to have less Maybes. I will only have one if getting
something form the store but not in the records I work with.
It will also allow me to change the logic of the relations later. An
implementation with a graph would not have the relations stored this way.
So it's an implementation detail.